+++ packages/dust_db_sqlite3/lib/dust_db_sqlite3.dart
@@ -1,4 +1,4 @@
-/// sqlite3 runtime for Dust DB.
+/// sqlite3 runtime for DustDB.
 library;
 
 export 'src/sqlite_pool.dart';